Vavr library, Actor & Entity Type declaration
Vavr.io is a functional designing library which will be very useful for tidying up a lot of the codebase and making development a little more fluent to work with.

Actor class now represents an entity itself, however the entity class is just for the more core relations like index id, such.. there's more to it but that's a simple way to explain.